+++ tests/integration_tests/utils.py
@@ -63,10 +63,10 @@ def wait_for_port(port, host="127.0.0.1", timeout=40.0):
                 ) from ex
 
 
-def w3_wait_for_new_blocks(w3, n):
+def w3_wait_for_new_blocks(w3, n, sleep=0.5):
     begin_height = w3.eth.block_number
     while True:
-        time.sleep(0.5)
+        time.sleep(sleep)
         cur_height = w3.eth.block_number
         if cur_height - begin_height >= n:
             break